+++ example/lib/markers.dart
@@ -14,8 +14,10 @@ class Place {
 
 class _MarkersPageState extends State<MarkersPage> {
   final mapController = MapController();
-  late final statefulMapController =
-      StatefulMapController(mapController: mapController);
+
+  late final statefulMapController = StatefulMapController(
+    mapController: mapController,
+  );
   late final StreamSubscription<StatefulMapControllerStateChange> sub;
 
   final List<Place> places = [
@@ -33,12 +35,14 @@ class _MarkersPageState extends State<MarkersPage> {
       if (!_markersOnMap.contains(place)) {
         debugPrint("Adding marker ${place.name}");
         statefulMapController.addMarker(
-            name: place.name,
-            marker: Marker(
-                point: place.point,
-                builder: (BuildContext context) {
-                  return const Icon(Icons.location_on);
-                }));
+          name: place.name,
+          marker: Marker(
+            point: place.point,
+            builder: (BuildContext context) {
+              return const Icon(Icons.location_on);
+            },
+          ),
+        );
         _markersOnMap.add(place);
         return;
       }
@@ -62,7 +66,11 @@ class _MarkersPageState extends State<MarkersPage> {
             zoom: 11.0,
           ),
           children: [
-            statefulMapController.tileLayer!,
+            TileLayer(
+              urlTemplate: 'https://tile.openstreetmap.org/{z}/{x}/{y}.png',
+              subdomains: const ['a', 'b', 'c'],
+              userAgentPackageName: 'dev.fleaflet.flutter_map.example',
+            ),
             MarkerLayer(
               markers: statefulMapController.markers,
             ),