+++ lib/src/exceptions.dart
@@ -5,13 +5,9 @@ class MarkerException implements Exception {
 
   /// The error message
   final String message;
-}
 
-/// An exception for a not implemented feature
-class NotImplementedException implements Exception {
-  /// Default constructor
-  const NotImplementedException(this.message);
-
-  /// The error message
-  final String message;
+  @override
+  String toString() {
+    return "MarkerException: $message";
+  }
 }